Add basic YAML support. (#57)
Adds basic support for displaying YAML files. jless will check the file
extension of the input file to determine whether to use the JSON parser, or the
YAML parser, defaulting to JSON if the file extension is unrecognized, or if
input comes from stdin. The input format can also be explicitly specified using
--yaml or --json flags.

YAML supports additional functionality on top of JSON:

Most notably, YAML supports non-scalar keys of mappings (i.e., Objects). Adding
support for exploring and focusing and collapsing individual components of
non-scalar keys would require significant changes. For now, we will treat the
JSON serialized version of the non-scalar key as a single string key, but we
will display that key using square brackets instead of quotes. These changes
will be in a separate commit.

YAML has anchors and aliases, which can be used to avoid repeated data in a
document. The YAML parser that we use automatically resolves aliases, and
throws an error on invalid ones. The "merge" operator ("<<:"), is not supported
by our parser, nor do we take steps to implement it ourselves.

YAML supports multiline strings. We replace any newlines with escaped newlines
(\n) for now. Adding proper multiline/line-wrap support would also require
significant changes throughout the codebase. It would be great if escaped
characters in string literals were highlighted differently (my color scheme
uses green for strings and orange for escaped characters). This would be useful
for YAML multiline strings, and for general escapes in JSON strings.

Since we convert the YAML documents to our own internal, lossy (due to no
support for non-scalar keys) format, we do not support any "pretty-printing" of
a YAML document when output is not a TTY.